Protect user modifying from common users

+++ src/app/api/users/[id]/route.ts
@@ -1,12 +1,26 @@
 /* eslint-disable @next/next/no-server-import-in-page */
 import { deleteUserById, updateUserRoleById } from '@repositories/user'
+import { getServerSession } from 'next-auth'
 import type { NextRequest } from 'next/server'
 import { NextResponse } from 'next/server'
+import { authOptions } from 'pages/api/auth/[...nextauth]'
 
 export async function PATCH(
     request: NextRequest,
     { params }: { params: { id: string } }
 ) {
+    const session = await getServerSession(authOptions)
+
+    if (!session) {
+        return new Response('Unauthorized', { status: 401 })
+    }
+
+    const sessionRole = session.user.role
+
+    if (sessionRole.toLocaleLowerCase() !== 'admin') {
+        return new Response('Unauthorized', { status: 401 })
+    }
+
     const id = params.id
     const { role } = await request.json()
 
@@ -29,6 +43,18 @@ export async function DELETE(
     request: NextRequest,
     { params }: { params: { id: string } }
 ) {
+    const session = await getServerSession(authOptions)
+
+    if (!session) {
+        return new Response('Unauthorized', { status: 401 })
+    }
+
+    const sessionRole = session.user.role
+
+    if (sessionRole.toLocaleLowerCase() !== 'admin') {
+        return new Response('Unauthorized', { status: 401 })
+    }
+
     const id = params.id
 
     const deleted = await deleteUserById(id)
